+++ CHANGELOG.md
@@ -30,6 +30,9 @@ ambiguous situations.
 
 ## Fixed
 - Crawling of exercises with instructions
+- Don't download unavailable elements.  
+  Elements that are unavailable (for example, because their availability is
+  time restricted) will not download the HTML for the info page anymore.
 
 ## 3.8.2 - 2025-04-29